LayoutTests/platform/mac-sonoma-wk2-lbse-text/svg/custom/focus-event-handling-expected.txt

TEST
Test whether focusin and focusout events are dispatched and seen in the focusin/focusout event handlers:

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL focusinSeen should be rect1. Was .
FAIL focusoutSeen should be rect1. Was .
PASS focusinSeen is "g"
PASS focusoutSeen is "g"
PASS focusinSeen is "use"
PASS focusoutSeen is "use"
PASS focusinSeen is "usesymbol"
PASS focusoutSeen is "usesymbol"
PASS focusinSeen is "switch"
PASS focusoutSeen is "switch"
PASS focusinSeen is "img"
PASS focusoutSeen is "img"
